// Kulthia Dasboard/src/app/core/models/opening-stock.model.ts
export interface OpeningStockModel {
    ProdCat: string;
    ProdNo: any;
    OldProdNo: any;
    Size:any;
    Pcs: any;
    GOLD: any;
    GrosWt:any;

    Making:  any;
    KundanMaking: any;
    Jobber:any;
    JobNo: any;
    Hallmarked: any;
    Certified:any;

    Kundan: any;
    JobberCharge: any;
    usersDetails:any;
}